Upgraded to Zustand

+++ src/components/task-list/task-item/task-item.tsx
@@ -2,7 +2,9 @@ import { MdOutlineDelete, MdOutlineEdit } from 'react-icons/md'
 
 import { type Task } from '../../../types'
 import style from './task-item.module.scss'
-import { useTaskStore } from '../../../features/todo-feature/todo-store'
+// import { useTaskStore } from '../../../features/todo-feature/todo-store'
+import { deleteTask, toggleTaskAsCompleted, selectTaskToEdit } from '../../../features/todo-feature/todo-store'
+
 
 type TaskItem = {
     data: Task
@@ -10,16 +12,19 @@ type TaskItem = {
 
 const TaskItem = ({ data }: TaskItem) => {
 
-    const { deleteTask, toggleTaskAsCompleted, selectTaskToEdit } = useTaskStore(
-        (state) => ({
-            deleteTask: state.deleteTask,
-            toggleTaskAsCompleted: state.toggleTaskAsCompleted,
-            selectTaskToEdit: state.selectTaskToEdit,
-        })
-    )
+    // const { deleteTask, toggleTaskAsCompleted, selectTaskToEdit } = useTaskStore(
+    //     (state) => ({
+    //         deleteTask: state.deleteTask,
+    //         toggleTaskAsCompleted: state.toggleTaskAsCompleted,
+    //         selectTaskToEdit: state.selectTaskToEdit,
+    //     })
+    // )
 
     const onDeleteTask = () => {
-        deleteTask(data.taskId)
+        const deleteTaskConfirm = confirm('Delete this task?')
+        if (deleteTaskConfirm) {
+            deleteTask(data.taskId)
+        }
     }
 
     const onSelectEditTask = () => {
